+++ tests/Traits/Visuals/ComponentVisualsTest.php
@@ -59,9 +59,7 @@ class ComponentVisualsTest extends TestCase
         } catch (ViewException $ViewException) {
             $this->testErrors = true;
 
-            // Temporary swapping to check exception throwing
-            //$this->assertSame('You must set a primary key using setPrimaryKey in the configure method.', substr($ViewException->getMessage(), 0, 71));
-            $this->assertSame('Typed property Rappasoft\LaravelLivewireTables\DataTableComponent::$pri', substr($ViewException->getMessage(), 0, 71));
+            $this->assertSame('You must set a primary key using setPrimaryKey in the configure method.', substr($ViewException->getMessage(), 0, 71));
 
         } catch (Exception $standardException) {
             $this->testErrors = true;